refactor and dividing into modules
