+++ .env
@@ -1,6 +1,6 @@
 VITE_APP_ENVIRONMENT=development
 VITE_APP_PORT=8096
-VITE_APP_API_URL='http://192.168.1.166:8000/_/api/'
+VITE_APP_API_URL='http://localhost:8000/_/api/'
 VITE_APP_APP_NAME='Web-Client NFT Books'
 VITE_APP_DEPLOY_ENVIRONMENT=''
 VITE_APP_FACEBOOK_LINK='https://www.facebook.com/distributedlab/'
