+++ src/common/image-editor/ImageEditor.vue
@@ -1,17 +1,34 @@
 <template>
   <div class="image-editor">
-    <div ref="editorContainerRef" class="image-editor__canvas-wrapper">
+    <loader v-if="!isEditorInitialized" />
+
+    <div
+      ref="editorContainerRef"
+      class="image-editor__canvas-wrapper"
+      :class="{
+        'image-editor__canvas-wrapper--hidden':
+          !isEditorInitialized || isInitFailed,
+      }"
+    >
       <canvas ref="editorCanvasRef" class="image-editor__canvas" />
     </div>
-    <image-editor-tool-kit />
+    <image-editor-tool-kit v-if="isEditorInitialized && !isInitFailed" />
+
+    <error-message
+      v-if="isInitFailed"
+      :title="$t('image-editor.error-message')"
+      :message="errorMsg"
+      :icon-name="$icons.photograph"
+    />
   </div>
 </template>
 
 <script setup lang="ts">
 import { ref, onMounted, provide } from 'vue'
-import { useImageEditor } from './composables'
-import { ImageEditorToolKit } from '.'
-import { EditorInstanceKey } from './types'
+import { Loader, ErrorMessage } from '@/common'
+import { useImageEditor } from '@image-editor/composables'
+import { ImageEditorToolKit } from '@image-editor'
+import { EditorInstanceKey } from '@image-editor/types'
 
 const props = withDefaults(
   defineProps<{
@@ -22,20 +39,29 @@ const props = withDefaults(
   },
 )
 
+const isEditorInitialized = ref(false)
+const isInitFailed = ref(false)
+const errorMsg = ref('')
 const editorContainerRef = ref<HTMLElement | null>(null)
 const editorCanvasRef = ref<HTMLCanvasElement | null>(null)
 
-const editorInstance = useImageEditor(editorCanvasRef, editorContainerRef)
-
-// providing canvas instance to all nested layers to avoid props drilling
-provide(EditorInstanceKey, { instance: editorInstance })
-
-const { init } = editorInstance
-
-onMounted(() => {
+onMounted(async () => {
   if (!editorCanvasRef.value || !editorContainerRef.value) return
 
-  init(props.imageUrl)
+  const editorInstance = useImageEditor(editorCanvasRef, editorContainerRef)
+
+  // providing canvas instance to all nested layers to avoid props drilling
+  provide(EditorInstanceKey, { instance: editorInstance })
+
+  try {
+    await editorInstance.init(props.imageUrl)
+  } catch (error) {
+    isInitFailed.value = true
+
+    if (error instanceof Error) errorMsg.value = error.message
+  }
+
+  isEditorInitialized.value = true
 })
 </script>
 
@@ -64,8 +90,12 @@ onMounted(() => {
   min-height: toRem(700);
   padding: toRem(40);
   background-color: var(--primary-light);
+  position: static;
 
-  // box-shadow: inset 0 0 toRem(18) toRem(-8) rgba(0, 0, 0, 1);
+  &--hidden {
+    visibility: hidden;
+    position: absolute;
+  }
 
   @include respond-to(small) {
     padding: toRem(20) 0;
