+++ src/common/image-editor/composables/index.ts
@@ -1,1 +1,5 @@
 export * from '@image-editor/composables/use-image-editor'
+export * from '@image-editor/composables/use-text'
+export * from '@image-editor/composables/use-drawing'
+export * from '@image-editor/composables/use-object-mutations'
+export * from '@image-editor/composables/use-canvas-operations'
