+++ src/forms/PurchaseBookForm.vue
@@ -31,7 +31,7 @@
     <modal v-model:is-shown="isEditingImage">
       <template #default="{ modal }">
         <div class="purchase-book-form__image-editor">
-          <image-editor />
+          <image-editor :image-url="book.banner.attributes.url" />
           <app-button :text="'CLOSE'" @click="modal.close" />
         </div>
       </template>
